Move map function to seperate file
